test(orchestration): split run-folder coverage by responsibility
